Refactor: atomic pattern에 따라 기존 main, footer 이동 - 디렉토리 변경
