+++ packages/wiii/frontend/components/organisms/HomeFooter.vue
